# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * IncidAI incident-management API
 * OpenAPI spec version: 0.1.0
 */
export type HealthStatusStatus = typeof HealthStatusStatus[keyof typeof HealthStatusStatus];


export const HealthStatusStatus = {
  ok: 'ok',
  degraded: 'degraded',
} as const;

export type HealthStatusDatabase = typeof HealthStatusDatabase[keyof typeof HealthStatusDatabase];


export const HealthStatusDatabase = {
  connected: 'connected',
  unavailable: 'unavailable',
} as const;

export type HealthStatusClassifier = typeof HealthStatusClassifier[keyof typeof HealthStatusClassifier];


export const HealthStatusClassifier = {
  ready: 'ready',
  unavailable: 'unavailable',
} as const;

export interface HealthStatus {
  status: HealthStatusStatus;
  database: HealthStatusDatabase;
  classifier: HealthStatusClassifier;
}

export type IncidentStatus = typeof IncidentStatus[keyof typeof IncidentStatus];


export const IncidentStatus = {
  open: 'open',
  in_progress: 'in_progress',
  resolved: 'resolved',
  escalated: 'escalated',
  closed: 'closed',
} as const;

export type IncidentPriority = typeof IncidentPriority[keyof typeof IncidentPriority];


export const IncidentPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface Incident {
  id: number;
  title: string;
  description: string;
  status: IncidentStatus;
  priority: IncidentPriority;
  /** @nullable */
  predictedCategory: string | null;
  /** @nullable */
  category: string | null;
  /**
     * @minimum 0
     * @maximum 1
     * @nullable
     */
  confidence: number | null;
  /** @nullable */
  reporterEmail: string | null;
  /** @nullable */
  assignedTo: string | null;
  /** @nullable */
  resolution: string | null;
  /** @nullable */
  resolutionMethod: string | null;
  resolutionVerified: boolean;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  resolvedAt: string | null;
}

export interface IncidentInput {
  /**
     * @minLength 3
     * @maxLength 240
     */
  title: string;
  /**
     * @minLength 8
     * @maxLength 10000
     */
  description: string;
  priority: IncidentPriority;
  reporterEmail?: string;
  /** @maxLength 160 */
  assignedTo?: string;
}

export interface IncidentUpdate {
  /**
     * @minLength 3
     * @maxLength 240
     */
  title?: string;
  /**
     * @minLength 8
     * @maxLength 10000
     */
  description?: string;
  status?: IncidentStatus;
  priority?: IncidentPriority;
  /**
     * @maxLength 160
     * @nullable
     */
  category?: string | null;
  /**
     * @maxLength 160
     * @nullable
     */
  assignedTo?: string | null;
}

export interface IncidentList {
  results: Incident[];
  total: number;
  limit: number;
  offset: number;
}

export interface AnalyzeIncidentInput {
  /**
     * @minimum 1
     * @maximum 5
     */
  similarLimit?: number;
}

export interface SimilarIncident {
  title: string;
  category: string;
  /** @nullable */
  resolution: string | null;
  /**
     * @minimum 0
     * @maximum 1
     */
  similarity: number;
}

export interface IncidentAnalysis {
  incident: Incident;
  predictedCategory: string;
  /**
     * @minimum 0
     * @maximum 1
     */
  confidence: number;
  method: string;
  rationale: string;
  similarIncidents: SimilarIncident[];
  recommendedResolution: string;
  resolutionMethod: string;
}

export interface ResolveIncidentInput {
  /**
     * @minLength 8
     * @maxLength 10000
     */
  resolution: string;
  verified: boolean;
}

export interface IncidentEvent {
  id: number;
  incidentId: number;
  eventType: string;
  message: string;
  createdAt: string;
}

export interface Category {
  name: string;
  historicalCount: number;
}

export interface CountByLabel {
  label: string;
  count: number;
}

export interface DashboardSummary {
  total: number;
  open: number;
  inProgress: number;
  resolved: number;
  escalated: number;
  byPriority: CountByLabel[];
  byCategory: CountByLabel[];
  recentIncidents: Incident[];
}

export type ListIncidentsParams = {
search?: string;
status?: IncidentStatus;
priority?: IncidentPriority;
category?: string;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
/**
 * @minimum 0
 */
offset?: number;
};