# lib/api-zod/src/generated/types/countByLabel.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * IncidAI incident-management API
 * OpenAPI spec version: 0.1.0
 */

export interface CountByLabel {
  label: string;
  count: number;
}