# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * IncidAI incident-management API
 * OpenAPI spec version: 0.1.0
 */
import type { CountByLabel } from './countByLabel';
import type { Incident } from './incident';

export interface DashboardSummary {
  total: number;
  open: number;
  inProgress: number;
  resolved: number;
  escalated: number;
  byPriority: CountByLabel[];
  byCategory: CountByLabel[];
  recentIncidents: Incident[];
}