# lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * IncidAI incident-management API
 * OpenAPI spec version: 0.1.0
 */
import type { HealthStatusClassifier } from './healthStatusClassifier';
import type { HealthStatusDatabase } from './healthStatusDatabase';
import type { HealthStatusStatus } from './healthStatusStatus';

export interface HealthStatus {
  status: HealthStatusStatus;
  database: HealthStatusDatabase;
  classifier: HealthStatusClassifier;
}