# lib/api-zod/src/generated/types/incidentAnalysis.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * IncidAI incident-management API
 * OpenAPI spec version: 0.1.0
 */
import type { Incident } from './incident';
import type { SimilarIncident } from './similarIncident';

export interface IncidentAnalysis {
  incident: Incident;
  predictedCategory: string;
  /**
     * @minimum 0
     * @maximum 1
     */
  confidence: number;
  method: string;
  rationale: string;
  similarIncidents: SimilarIncident[];
  recommendedResolution: string;
  resolutionMethod: string;
}