# lib/api-zod/src/generated/types/incidentStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * IncidAI incident-management API
 * OpenAPI spec version: 0.1.0
 */

export type IncidentStatus = typeof IncidentStatus[keyof typeof IncidentStatus];


export const IncidentStatus = {
  open: 'open',
  in_progress: 'in_progress',
  resolved: 'resolved',
  escalated: 'escalated',
  closed: 'closed',
} as const;